Moves request logic into wizard
