+++ frontend/src/utils/complianceRequestWizard.ts
@@ -1,4 +1,6 @@
- export type Mode = 'new' | 'edit' | 'view';
+import { type ComplianceFormDataResponse, type ComplianceRequestResponse } from '../api/types';
+
+export type Mode = 'new' | 'edit' | 'view';
 
 export interface FormState {
   csip_aus_version: string;
@@ -30,4 +32,50 @@ export function emptyForm(): FormState {
     software_client_versions: '',
     onsite_hardware_details: '',
   };
+}
+
+export function buildInitialForm(
+  formData: ComplianceFormDataResponse,
+  prefill: ComplianceRequestResponse | undefined,
+  opts: { prefillClasses: boolean; prefillRuns: boolean }
+): FormState {
+  const form = emptyForm();
+  form.csip_aus_version = prefill?.csip_aus_version || formData.csipaus_versions[0] || '';
+  if (prefill) {
+    form.witnessed_at = prefill.witnessed_at.split('T')[0];
+    form.der_brand = prefill.der_brand;
+    form.der_oem = prefill.der_oem;
+    form.der_series = prefill.der_series;
+    form.der_representative_models = prefill.der_representative_models;
+    form.software_client_type = prefill.software_client_type || 'direct';
+    form.software_client_providers = prefill.software_client_providers;
+    form.software_client_versions = prefill.software_client_versions;
+    form.onsite_hardware_details = prefill.onsite_hardware_details;
+  }
+
+  const classMap = formData.tests_by_version_and_class[form.csip_aus_version] ?? {};
+  const completed = new Set(formData.completed_test_procedures);
+
+  if (prefill && opts.prefillClasses) {
+    form.classes = new Set(prefill.classes);
+  } else {
+    // New request: preselect classes whose required tests all have a successful run.
+    form.classes = new Set(
+      Object.keys(classMap).filter((c) => (classMap[c] ?? []).every((p) => completed.has(p)))
+    );
+  }
+
+  // Default each procedure's run to its first successful run, then apply any prefilled selections.
+  const runs = formData.successful_runs;
+  for (const run of runs) {
+    form.runByProcedure[run.test_procedure_id] ??= run.run_id;
+  }
+  if (prefill && opts.prefillRuns) {
+    const procedureByRun = new Map(runs.map((r) => [r.run_id, r.test_procedure_id]));
+    for (const runId of prefill.runs) {
+      const procedure = procedureByRun.get(runId);
+      if (procedure) form.runByProcedure[procedure] = runId;
+    }
+  }
+  return form;
 }
